Split fieldValue.test.js as it was too long
